\name{moduleGene}
\alias{moduleGene}
\alias{modulegene}

\title{
	Functional module gene function.
}

\description{
	Based on the windowed Seurat object file, explore functional gene modules in spatial locations using clustering and WGCNA methods.
}

\usage{
	moduleGene(rds, rdsList = FALSE, dims = 10, resolution = 0.8, geneVector = NULL, ColumnCluster = FALSE, strict = TRUE, show_rownames = TRUE, show_colnames = TRUE, angle_col = "45", width = 18, height = 18, sv = 'V3')
}

\arguments{
  \item{rds}{       
	The Seurat object file processed using wcellDue from the R package SpLin.
  }
  \item{rdsList}{       
	Whether it is a list of Seurat objects, where each Seurat object is generated by the function wcellDue from the R package SpLin.
  }
  \item{dims}{       
	Dimensions of reduction to use as input.
  }
  \item{geneVector}{       
	A vector composed of gene names, used for gene screening in gene pattern analysis. The default value is empty.
  }
  \item{resolution}{       
	Value of the resolution parameter, use a value above(below) 1.0 if you want to obtain a larger (smaller) number of communities.
  }  
  \item{ColumnCluster}{       
	boolean values determining if columns should be clustered or 'hclust' object.
  } 
  \item{strict}{       
	Whether to filter out non-differentially expressed genes using FindAllMarkers.
  }   
  \item{show_rownames}{       
	boolean specifying if column names are be shown.
  }   
  \item{show_colnames}{       
	boolean specifying if column names are be shown.
  }
  \item{angle_col}{       
	angle of the column labels, right now one can choose only from few predefined options (0, 45, 90, 270 and 315).
  }
  \item{width}{       
	The width of the PDF figure.
  }  
  \item{height}{       
	The height of the PDF figure.
  } 
  \item{sv}{       
	The Seurat object of class, V5 assay or V3 assay. Generally, versions of Seurat prior to V5 (excluding V5) are considered V3 assay. The parameter value can be either V3 or V5, and it defaults to V3 here.
  }      
}
\author{
	Wen Luo(design, implementation, testing)
}
\examples{
	rds = 'HE/IGSI/winCell/windowCell.RDS'
	dims <- 10
	resolution <- 0.8
	strict <- TRUE
	show_rownames = TRUE
	show_colnames = TRUE
	angle_col = "45"   
	width = 28
	height = 42
	moduleGene(rds, dims = dims, resolution = resolution, geneVector = NULL, strict = strict, show_rownames = show_rownames, show_colnames = show_colnames, angle_col = angle_col, width = width, height = height)
}