\name{pixelEXPPlot}
\alias{pixelEXPPlot}
\alias{pixelexpplot}

\title{
	Visualization of gene expression levels and pixel points after straightening.
}

\description{
	Used for visualizing the distribution of gene expression levels in the straightened coordinate system.
}

\usage{
	pixelEXPPlot(rds, pixelPointsFile, markerGenes = c('Reg3b','Dcn','Hmgcs2'), imageType = 'HE', tolerance = 10, xmin = NULL, xmax = NULL, SpatialFeaturePlotstriptextsize = 20, SpatialFeaturePlotpointsize = 0.5, SpatialFeaturePlotpointalpha = 1, sv = 'V3')
						
}

\arguments{
  \item{rds}{       
     The Seurat object file processed using the R package SpLin.
  }
  \item{pixelPointsFile}{       
     A text file generated by the R package SpLin, containing the color values of the pixel matrix, the original coordinate system, and the values of the straightened coordinate system. The file must have five columns with the column names "Cell", "x", "y", "bin", "xnew", and "ynew".
  }  
  \item{markerGenes}{       
     A list of genes of interest, with the input format being a vector.
  }
  \item{imageType}{       
     Type of the original image, can be 'ssDNA' or 'HE'.
  }
  \item{tolerance}{       
     A range of tolerance that defines how close a color is to black or white. It allows colors within a certain range of black or white to be considered close to black or white.
  }    
  \item{xmin}{       
     The minimum value of the x-coordinate for the region to be visualized.
  }
  \item{xmax}{       
     The maximum value of the x-coordinate for the region to be visualized.
  }  
  \item{SpatialFeaturePlotstriptextsize}{       
     Color bar text size.
  }
  \item{SpatialFeaturePlotpointsize}{       
     The size of the scatter points.
  }
  \item{SpatialFeaturePlotpointalpha}{       
     The transparency of the scatter points, ranging from 0 to 1.
  }
  \item{sv}{       
	The Seurat object of class, V5 assay or V3 assay. Generally, versions of Seurat prior to V5 (excluding V5) are considered V3 assay. The parameter value can be either V3 or V5, and it defaults to V3 here.
  }    
}
\author{
	Wen Luo(design, implementation, testing)
}
\examples{
	# 1
	rds = 'HE/SpLin_EXP_output.rds'
	pixelPointsFile = 'HE/pixelPointsUpdate.txt'  
	markerGenes = c('Muc2')
	xmin = NULL
	xmax = NULL 
	SpatialFeaturePlotstriptextsize = 20
	SpatialFeaturePlotpointsize = 1.5
	SpatialFeaturePlotpointalpha = 1
	library(SpLin)
	pixelEXPPlot(rds, pixelPointsFile, markerGenes = markerGenes, xmin = xmin, xmax = xmax, SpatialFeaturePlotstriptextsize = SpatialFeaturePlotstriptextsize, SpatialFeaturePlotpointsize = SpatialFeaturePlotpointsize, SpatialFeaturePlotpointalpha = SpatialFeaturePlotpointalpha)
	
	# 2
	SpatialFeaturePlotstriptextsize = 20
	SpatialFeaturePlotpointsize = 3
	SpatialFeaturePlotpointalpha = 1
	xmin = 0
	xmax = 5000
	pixelEXPPlot(rds, pixelPointsFile, markerGenes = markerGenes, xmin = xmin, xmax = xmax, SpatialFeaturePlotstriptextsize = SpatialFeaturePlotstriptextsize, SpatialFeaturePlotpointsize = SpatialFeaturePlotpointsize, SpatialFeaturePlotpointalpha = SpatialFeaturePlotpointalpha)
}