\name{preAutoRegister}
\alias{preAutoRegister}
\alias{preautoregister}

\title{
	Generate the mask image.
}

\description{
	Since this function is used to generate a mask image for registration operations, subsequent analyses will rely on the shape of the mask image to determine whether a mirror transformation has occurred.
}

\usage{
	preAutoRegister(json, cellPointsFile, idjson, Signal = FALSE)
}

\arguments{
  \item{json}{       
     This is a Labelme JSON result file for images containing the labels "Nei" and "Wai", or additionally including the "Sign" label after adjustment.
  }
  \item{cellPointsFile}{       
     A text file generated by the R package SpLin, containing a matrix without row or column names, with values derived from points in the signal matrix. The file is typically named "cellPoints.txt".
  }
  \item{idjson}{       
     The dimension information file for the image, typically named "Dimensions.txt".
  }
  \item{Signal}{       
     Whether it is an image formed by the signal matrix.
  }
}
\author{
	Wen Luo(design, implementation, testing)
}
\examples{
	json = 'SignalMatrix/Adjusted_signalMatrix.json'
	cellPointsFile = 'SignalMatrix/cellPoints.txt'
	idjson = 'SignalMatrix/image_dimensions.json' 
	library(SpLin)
	preAutoRegister(json, cellPointsFile, idjson, Signal = TRUE)
}