\name{preAutoRegisterHE}
\alias{preAutoRegisterHE}
\alias{preautoregisterhe}

\title{
	Generate the mask image.
}

\description{
	Since this function is used to generate a mask image for registration operations, subsequent analyses will rely on the shape of the mask image to determine whether a mirror transformation has occurred.
}

\usage{
	preAutoRegisterHE(markingPointsFile, cellPointsFile, idjson)
}

\arguments{
  \item{markingPointsFile}{       
    A text file generated by the R package SpLin, containing a matrix with column names (X, Y, Color) but no row names. The values are derived from the points in the HE pixel matrix. The file name is typically Adjusted_output_coordinates_and_colors.txt.
  }
  \item{cellPointsFile}{       
     A text file generated by the R package SpLin, containing a matrix with column names (X, Y, Color) but no row names. The values are derived from the points in the ssDNA pixel matrix. The file name is typically Adjusted_output_coordinates_and_colors.txt.
  }
  \item{idjson}{       
     The dimension information file for the image, typically named "Dimensions.txt".
  }
}
\author{
	Wen Luo(design, implementation, testing)
}
\examples{
	markingPointsFile = 'HE/Adjusted_output_coordinates_and_colors.txt'
	cellPointsFile = 'ssDNA/Adjusted_output_coordinates_and_colors.txt'
	idjson = 'ssDNA/image_dimensions.json'
	library(SpLin)
	preAutoRegisterHE(markingPointsFile, cellPointsFile, idjson)
}