\name{RapidRegisterHE}
\alias{RapidRegisterHE}
\alias{rapidregisterhe}

\title{
	Manual registration function.
}

\description{
	This manual verification-based registration method operates through the following workflow: (a) performing initial alignment by setting iterative points according to mask contour morphology, (b) selecting optimal parameters for refinement based on alignment results, and (c) generating the final registered solution.
}

\usage{
	RapidRegisterHE(pixelssDNAFile, pixelHEFile, markingPointsImage, markingPoints_MASKJSON, cellPointsImage, cellPoints_MASKJSON, idjson, transform = NULL, kpixel = 10, epsilon = 0.001, interval = 0.05, intervalAngle = 30, up = 0, down = 0, left = 0, right = 0, theta = 0)
}

\arguments{
  \item{pixelssDNAFile}{       
     A text file generated by the R package SpLin, containing a matrix with column names (X, Y, Color) but no row names. The values are derived from the points in the ssDNA pixel matrix. The file name is typically Adjusted_output_coordinates_and_colors.txt.
  }  
  \item{pixelHEFile}{       
     A text file generated by the R package SpLin, containing a matrix with column names (X, Y, Color) but no row names. The values are derived from the points in the HE pixel matrix. The file name is typically Adjusted_output_coordinates_and_colors.txt.
  }  
  \item{markingPointsImage}{       
     The mask image of the HE (Hematoxylin and Eosin) pixel matrix.
  }  
  \item{markingPoints_MASKJSON}{       
     The Labelme JSON file for the initial iteration points of the mask image of the HE (Hematoxylin and Eosin) pixel matrix.
  }  
  \item{cellPointsImage}{       
     The mask image of the ssDNA pixel matrix.
  }  
  \item{cellPoints_MASKJSON}{       
     The Labelme JSON file for the initial iteration points of the mask image of the ssDNA pixel matrix.
  }   
  \item{idjson}{       
     The dimension information file for the image, typically named "Dimensions.txt".
  }  
  \item{transform}{       
     If the contour of the annotated points is in a mirror relationship with the contour of the expression matrix, adjust the relationship of the annotated points' contour based on the contour of the expression matrix. The parameter can be "horizontal," "vertical," or NULL.
  }  
  \item{kpixel}{       
     The pixel interval, used to evaluate the registration coefficient, defaults to 10.
  }  
  \item{epsilon}{       
     The error ratio caused by manual Lasso and annotation, which defaults to 0.001.
  }      
  \item{interval}{       
     The radius interval of concentric circles.
  }   
  \item{intervalAngle}{       
     The angular interval, used for drawing lines at each specified angle.
  }   
  \item{up}{       
     The positive direction of the Y-axis, where positive values indicate stretching outward from the concentric circles, and negative values indicate scaling inward towards the concentric circles.
  }   
  \item{down}{       
     The negative direction of the Y-axis, where positive values indicate stretching outward from the concentric circles, and negative values indicate scaling inward towards the concentric circles.
  }   
  \item{left}{       
     The negative direction of the X-axis, where positive values indicate stretching outward from the concentric circles, and negative values indicate scaling inward towards the concentric circles.
  }   
  \item{right}{       
     The positive direction of the X-axis, where positive values indicate stretching outward from the concentric circles, and negative values indicate scaling inward towards the concentric circles.
  }   
  \item{theta}{       
     The rotation angle, where positive values indicate counterclockwise rotation, and negative values indicate clockwise rotation.
  }
}
\author{
	Wen Luo(design, implementation, testing)
}
\examples{
	pixelssDNAFile = 'ssDNA/Adjusted_output_coordinates_and_colors.txt'
	pixelHEFile = 'HE/Adjusted_output_coordinates_and_colors.txt'
	markingPointsImage = 'HE/HE_pixel_MASK.png'
	markingPoints_MASKJSON = 'HE/HE_pixel_MASK.json'
	cellPointsImage = 'HE/ssDNA_pixel_MASK.png'
	cellPoints_MASKJSON = 'HE/ssDNA_pixel_MASK.json'
	idjson = 'ssDNA/image_dimensions.json'
	transform = NULL
	kpixel = 10
	epsilon = 0.001
	interval = 0.05
	intervalAngle = 30
	up = 0
	down = 0
	left = 0
	right = 0
	theta = 0
	library(SpLin)
	RapidRegisterHE(pixelssDNAFile, pixelHEFile, markingPointsImage, markingPoints_MASKJSON, cellPointsImage, cellPoints_MASKJSON, idjson, transform = transform, kpixel = kpixel, epsilon = epsilon, interval = interval, intervalAngle = intervalAngle, up = up, down = down, left = left, right = right, theta = theta)
}