\name{RapidRegisterOUT}
\alias{RapidRegisterOUT}
\alias{rapidregisterout}

\title{
	Registration result extraction function.
}

\description{
	This function, based on the registered results, extracts the final registered information from the results using the provided index value and generates the result data and its files.
}

\usage{
	RapidRegisterOUT(rds, json, idjson, pixelssDNAFile = NULL, pixelHEFile = NULL, index = 1)
}

\arguments{
  \item{rds}{       
     The Seurat object file.
  }
  \item{json}{       
     The JSON file processed using the R package SpLin.
  }
  \item{idjson}{       
     The dimension information file for the image, typically named "Dimensions.txt".
  }  
  \item{pixelssDNAFile}{       
     A text file generated by the R package SpLin, containing a matrix with column names (X, Y, Color) but no row names. The values are derived from the points in the ssDNA pixel matrix. The file name is typically Adjusted_output_coordinates_and_colors.txt.
  }  
  \item{pixelHEFile}{       
     A text file generated by the R package SpLin, containing a matrix with column names (X, Y, Color) but no row names. The values are derived from the points in the HE pixel matrix. The file name is typically Adjusted_output_coordinates_and_colors.txt.
  }  
  \item{index}{       
     Which index corresponds to the dataset to be selected.
  }
}
\author{
	Wen Luo(design, implementation, testing)
}
\examples{
	# 1
	rds = 'modified_data.rds'
	json = 'SignalMatrix/AllRegistrationSchemes.json'
	idjson = 'SignalMatrix/image_dimensions.json'
	pixelssDNAFile = NULL
	pixelHEFile = NULL
	index = 2
	library(SpLin)
	RapidRegisterOUT(rds, json, idjson, pixelssDNAFile = pixelssDNAFile, pixelHEFile = pixelHEFile, index = index)
	
	# 2
	rds = 'SpLin_input.rds'
	json = 'ssDNA/AllRegistrationSchemes.json'
	idjson = 'ssDNA/image_dimensions.json'
	pixelssDNAFile = 'ssDNA/Adjusted_output_coordinates_and_colors.txt' 
	pixelHEFile = 'HE/Scheme/1/ssDNAHE_Adjusted_output_coordinates_and_colors.txt' 
	index = 1
	library(SpLin)
	RapidRegisterOUT(rds, json, idjson, pixelssDNAFile = pixelssDNAFile, pixelHEFile = pixelHEFile, index = index)	
}