\name{SpLin.help}
\alias{SpLin.readme}
\alias{readme.SpLin}
\alias{splin.help}
\alias{help.SpLin}
\alias{SpLin}
\alias{splin}


\title{
	Welcome to SpLin!
}

\description{

SpatioLinear (SpLin) is an R package designed to process spatiotemporal omics data, particularly from the chip of spatial transcriptomic. It accurately linearizes and horizontally maps the physical coordinates of specific curved-shaped, continuous local tissue structures (e.g., neural tubes, intestinal tissues) to the x-axis. By algorithmically transforming these complex tissue architectures into a standardized 1D coordinate system, SpatioLinear preserves topological integrity while enabling more intuitive display and analysis of the spatial distribution features of these tissue structures.

//////////////////////////////////////////////// \cr
https://github.com/scholarLW/SpLin \cr
////////////////////////////////////////////////

It is highly recommended that you visit the webpage for this project as it contains the most recently updated
information and instructions regarding everything patchwork.

}

%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%}

%\references{
%% ~put references to the literature/web site here ~
%}

\author{
Wen Luo(design, implementation, testing), \email{luokmhd@foxmail.com}
}
%\note{
%%  ~~further notes~~
%}
